+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Fix setting client in UploadLogResponse (Fix [#9](https://github.com/aternosorg/mclogs-java/issues/9))
+- Add Path based variants to listCrashReportsInDirectory and listLogsInDirectory